Crate democracy_mbm

Re-exports

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

versioned_migrations

Exports for versioned migration types for this pallet.

weights

Autogenerated weights for democracy_mbm

Structs

DemocracyMigrationSaveMigrationBlock

DemocracyMigrationV1ToV2

MigrationState

Stores the migration step and the block number at which the migration started

Enums

MigrationStep

Progressive migration state to keep track of progress